requirements.txt
openai
python-dotenv
pyyaml
rich
prompt_toolkit
mcp
websockets
